Hibernate property hibernate.hbm2ddl.auto is added
hibernate.hbm2.ddl.auto is set to "create" . This facilitates automatic
creation of tables in the db corresponding to the model classes.  If
tables already exist in the table, they will be dropped and recreated.
This property can be set to update if dropping and recreating
existing tables are not needed