Spring web flow configuration and spring security using jdbc authentication
Spring web flow initial configuration. The base path is set to /WEB-INF.
The path b/w the base path and flow.xml determines the id of the flow.
The request for http://localhost:8080/<project-name>/flows starts the
flow and enters the first state defined in flow.xml.
Subsequent states are triggered by actions specified in the input tag
with name prefixed with _event_id in the corresponding html/jsp page.
The ${flowexectionUrl} variable needs to be used in the action tag with
method="post"

Authentication and authorization is done through spring security's jdbc
authentication provider using custom query and tables instead of default
schema(users and authorities)