Save Product functionality implemented
Binding  nested objects Category and Supplier inside product.
Supplier and Category options are selected from the select dropdown
list.
The category and supplier properties are mapped appropriately in the
name/path attribute of the input/select tag
Ex. property 'name' of category is mapped as 'category.name' in the
name/path attribute of the input/select tag.