Display Order details in Order Confirmation page
Following method added in OrderDAO.
getOrdersByUser() which retrieves all the orders for a particular user.
This method can be used to retrieve all the orders for the currently
logged in user and display it in the order confirmation page.
Each order object in the retrieved orders also implicitly contains the
list of products because of hibernate mapping.