Guest user access to view products, home, about , contact us pages provided
Previously all users needed to be authenticated before accessing any of
the pages(home, products pages).

In this commit access has been provided for the guest user to view
products in the e-commerce website. He needs to be authenticated only
when he buys a product.

spring security tags library is used to control access to specific pages
using the <sec:authorize> tags

Spring security taglib, spring-core, spring-web versions changed to
4.2.x.  The following error is resolved after adding the above versions
of spring modules.
-no classdef
org.springframework.security.web.context.support.SecurityWebApplicationContextUtils

Spring security taglib declaration added after the <!DOCTYPE> The
<sec:authorize> tag being recognized by jsp after that change.

Logged in username is displayed using
${pageContext.request.userPrincipal.name}

ServletContext is used to store categories and suppliers in application
scope rather than in session scope.
When user logs out the session gets invalidated and the variables
categories and suppliers are also cleared. Hence they are stored in
application scope so that they are available even if session is
invalidated when the user logs out.